+++ packages/glyph-components/.storybook/main.js
@@ -13,6 +13,10 @@ module.exports = {
             use: [ 'style-loader', 'css-loader', 'sass-loader' ],
             include: path.resolve(__dirname, '../'),
         });
+        config.node = {
+            fs: 'empty',
+            global: true,
+        };
         config.plugins.push(new ProgressBarPlugin());
         // Return the altered config
         return config;