+++ packages/glyph-core/src/index.ts
@@ -4,4 +4,6 @@ import rootReducer from './reducers';
 export * from './utils';
 export * from './models';
 export * from './enums';
+export * from './middlewares';
+export * from './composer';
 export { actions, rootReducer };